Updating the HN Shoulder blocking script for a call to xml-based beamset files housed in protocols. -Added basic functionality for a plan copy to be made.
